% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two no. together}
\usage{
add(x, y)
}
\arguments{
\item{x}{a no.}

\item{y}{a no.}
}
\description{
Add 2 no. together
}
\details{
more detailed info provided here.
}
\examples{
add(1, 2)
add(1000, 1345)
}
\author{
Mads
}
